extract out some of the nafuda tooling functions into their own file
